pyproject.toml
[tool.poetry]
name = "HarukaBot"
version = "3.0.0-alpha.0"
description = "The multiple function QQ bot based on NoneBot."
authors = ["mnixry"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-alpha.6"
confuse = "^1.4.0"
sqlalchemy = "^1.3.20"
pillow = "^8.0.1"
lxml = "^4.6.2"
feedparser = "^6.0.2"
requests = "^2.25.0"
aiofiles = "^0.6.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
mypy = "^0.790"
isort = "^5.6.4"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
